change variable type from int to float for the inputs

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 income = input("Enter your monthly income:")
 expense = input("Enter your monthly expenses:")
-monthly_savings = int(income) - int(expense);
+monthly_savings = float(income) - float(expense);
 project_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 print(f"Your monthly savings are ${monthly_savings}")
 print(f"Projected savings after one year, with interest, is: ${project_savings}")
